backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
ultralytics 
pillow 
easyocr 
python-dotenv 
pymongo 
opencv-python-headless 
numpy 
utils